Sanergy/requirements.txt
asgiref==3.2.3
autopep8==1.5
beautifulsoup4==4.8.2
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
coverage==5.0.3
cryptography==2.8
defusedxml==0.6.0
Django==3.0.2
django-bootstrap3==12.0.3
django-bootstrap4==1.1.1
django-crispy-forms==1.8.1
django-extended-choices==1.3.3
django-pandas==0.6.1
django-salesforce==0.9
django-seed==0.1.9
django-sendgrid-v5==0.8.1
djangorestframework==3.11.0
Faker==4.0.0
flatten-json==0.1.7
future==0.18.2
idna==2.8
isort==4.3.21
json-flatten==0.1
jsonpickle==1.2
lazy-object-proxy==1.4.3
mccabe==0.6.1
nose==1.3.7
numpy==1.18.1
oauthlib==3.1.0
pandas==0.25.3
Pillow==7.0.0
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pycparser==2.19
PyJWT==1.7.1
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-http-client==3.2.4
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
sendgrid==6.1.1
simple-salesforce==0.74.3
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==1.9.5
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.25.7
